test: fix flaky test-readline-interface
Move test reliant on timer triggering in a timely fahion from parallel
to sequential. The test can fail under high load when the timer is
triggered too late and the `\r` and `\n` are treated as separate lines.

PR-URL: https://github.com/nodejs/node/pull/15066
