+++ test/parallel/test-readline-interface.js
@@ -291,30 +291,6 @@ function isWarned(emitter) {
     }), delay * 2);
   }
 
-  // Emit one line events when the delay between \r and \n is
-  // over the default crlfDelay but within the setting value
-  {
-    const fi = new FakeInput();
-    const delay = 125;
-    const crlfDelay = common.platformTimeout(1000);
-    const rli = new readline.Interface({
-      input: fi,
-      output: fi,
-      terminal: terminal,
-      crlfDelay
-    });
-    let callCount = 0;
-    rli.on('line', function(line) {
-      callCount++;
-    });
-    fi.emit('data', '\r');
-    setTimeout(common.mustCall(() => {
-      fi.emit('data', '\n');
-      assert.strictEqual(callCount, 1);
-      rli.close();
-    }), delay);
-  }
-
   // set crlfDelay to `Infinity` is allowed
   {
     const fi = new FakeInput();
